+++ lib/ueberauth/strategy/google/oauth.ex
@@ -27,10 +27,14 @@ defmodule Ueberauth.Strategy.Google.OAuth do
   """
   def client(opts \\ []) do
     config = Application.get_env(:ueberauth, __MODULE__, [])
-    opts = @defaults |> Keyword.merge(config) |> Keyword.merge(opts) |> resolve_values()
     json_library = Ueberauth.json_library()
 
-    OAuth2.Client.new(opts)
+    @defaults
+    |> Keyword.merge(config)
+    |> Keyword.merge(opts)
+    |> resolve_values()
+    |> generate_secret()
+    |> OAuth2.Client.new()
     |> OAuth2.Client.put_serializer("application/json", json_library)
   end
 
@@ -89,4 +93,14 @@ defmodule Ueberauth.Strategy.Google.OAuth do
 
   defp resolve_value({m, f, a}) when is_atom(m) and is_atom(f), do: apply(m, f, a)
   defp resolve_value(v), do: v
+
+  defp generate_secret(opts) do
+    if is_tuple(opts[:client_secret]) do
+      {module, fun} = opts[:client_secret]
+      secret = apply(module, fun, [opts])
+      Keyword.put(opts, :client_secret, secret)
+    else
+      opts
+    end
+  end
 end